Struct ledger_mob::tx::TransactionHandle

pub struct TransactionHandle<T: Device> { /* private fields */ }

Handle to a hardware wallet configured for transaction execution

See DeviceHandle::transaction to create a TransactionHandle

Implementations

impl<T: Device> TransactionHandle<T>

pub async fn key_image( &self, subaddress_index: u64, tx_out_public_key: &TxOutPublic ) -> Result<KeyImage, Error>

Asynchronously compute key image for a given subaddress and tx_out_public_key.

See [KeyImageComputer] for the public blocking API

impl<T: Device> TransactionHandle<T>

pub async fn memo_sign( &self, sender_subaddress_index: u64, tx_public_key: &TxOutPublic, target_subaddress: PublicSubaddress, memo_type: &[u8; 2], memo_data_sans_hmac: &[u8; 48] ) -> Result<[u8; 16], Error>

Asynchronously compute the HMAC signature for the provided memo and target address.

See [MemoHmacSigner] for the public blocking API.

impl<T: Device> TransactionHandle<T>

pub async fn ring_sign( &self, _message: &[u8], signable_ring: &SignableInputRing, pseudo_output_blinding: Scalar ) -> Result<RingMLSAG, Error>

Asynchronously execute a ring signing operation on ledger hardware.

See [RingSigner] trait for public / blocking API

impl<T: Device> TransactionHandle<T>

pub async fn view_subaddress(&self, index: u64) -> Result<ViewSubaddress, Error>

Asynchronously fetch a view subaddress by subaddress index, inheriting the account index from the transaction context.

See [Subaddress] trait for public (blocking) API

impl<T: Device + Send> TransactionHandle<T>

pub async fn set_tx_summary( &mut self, block_version: BlockVersion, message: &[u8], summary: &TxSummary, unblinding: &TxSummaryUnblindingData ) -> Result<(), Error>

Load tx summary for signing operation, alternative to set_message for block versions > 3

impl<T: Device + Send> TransactionHandle<T>

pub async fn new( info: TxConfig, transport: Arc<Mutex<T>> ) -> Result<Self, Error>

Initialise a new transaction over the provided device

pub async fn set_message(&mut self, m: &[u8]) -> Result<(), Error>

Set message for transaction

pub async fn await_approval(&mut self, timeout_s: u32) -> Result<(), Error>

Await on-device transaction approval

pub async fn complete(self) -> Result<(), Error>

Signal transaction completion